Write resetup file when old master is stuck
